refactor: 기존 룩북 페이지의 새로고침 로직을 커스텀 훅으로 분리
